refactor security audit workflow
